+++ rate.h
@@ -1,8 +1,8 @@
 using namespace std;
 class rate {
   public:
-    float Ra_prime(float Sa);
-    float Rb_prime(float Sb);
+    int Ra_prime(float Sa);
+    int Rb_prime(float Sb);
     float Ea();
     float Eb();
     void set_argument(int temp_K, float temp_Ra, float temp_Rb);
